remove duplicated mode change code
