+++ src/features/swap/components/Swap.tsx
@@ -9,12 +9,9 @@ import { Details } from './Details';
 import { PendingWithdrawals } from './PendingWithdrawals';
 import { SwapForm } from './SwapForm';
 
-interface SwapProps {
-  mode: Mode;
-  onModeChange: (mode: Mode) => void;
-}
 
-export const Swap = ({ mode, onModeChange }: SwapProps) => {
+
+export const Swap = ({ mode }: SwapProps) => {
   const { pendingWithdrawals } = useAccountContext();
   const { amount, setAmount, swap, balance, receiveAmount, swapRate, gasFee, setMaxAmount } =
     useSwap(mode);
@@ -32,7 +29,6 @@ export const Swap = ({ mode, onModeChange }: SwapProps) => {
         setMaxAmount={setMaxAmount}
         onSubmit={swap}
         onChange={setAmount}
-        onModeChange={onModeChange}
       />
       <OpacityTransition id={mode}>
         <div className="w-full px-[8px]">
