+++ src/pages/[...slug].tsx
@@ -8,13 +8,8 @@ const SwapPage: NextPage = () => {
   const { slug } = router.query as { slug?: string[] };
   const mode = slug ? slug[0] : Mode.stake;
 
-  const onModeChange = (mode: Mode) => {
-    void router.push({
-      pathname: `/${mode}`,
-    });
-  };
   if (typeof mode !== 'string' || !Object.values(Mode).includes(mode as Mode)) return null;
-  return <Swap mode={mode as Mode} onModeChange={onModeChange} />;
+  return <Swap mode={mode as Mode} />;
 };
 
 export default SwapPage;
