+++ src/pages/validators.tsx
@@ -1,16 +1,8 @@
 import type { NextPage } from 'next';
-import { useRouter } from 'next/router';
 import { Validators } from 'src/features/validators/components/Validators';
-import { Mode } from 'src/types';
 
 const ValidatorsPage: NextPage = () => {
-  const router = useRouter();
-  const onModeChange = (mode: Mode) => {
-    void router.push({
-      pathname: `/${mode}`,
-    });
-  };
-  return <Validators onModeChange={onModeChange} />;
+  return <Validators />;
 };
 
 export default ValidatorsPage;
